Improve type_info extraction: fall back to `resolved`
Summary:
The Pyre `typeCheck` code has a lot of accidental complexity because if its
history - for example, `TypeInfo.Unit.t` was added well after the beginning
of Pyre to handle refinement types, and many expression resolutions don't
set a value for it.

None of that complexity concerns the `locationBasedLookupProcessor` directly
though, there we really just want to use a combination of refined and non-refined
types to produce some coverage information for IDE, static analysis tools, and
type coverage reporting.

So, before we call into the `record_expression_type_info` hook, we should
ensure that a type info is present by falling back to a default type info
when only `Resolved.resolved` is present in the output of `forward_expression`.

This fixes both regressions identified in unit tests for D62449018:
- The `tuple[unknown, unknown]` goes back to `tuple[float, int]`
- The `x == 0` as `Any` goes back to `bool`

Reviewed By: grievejia

Differential Revision: D62449724

fbshipit-source-id: 4ca4f5074d53516e4d1cee730e563e3028020788