Rewire the types query to use proper recursion
Summary:
This commit is at the heart of the current stack.

It cuts Pyre's `locationBasedLookup` module to use the types
saved through the Context in `typeCheck` whenever possible.

Because `typeCheck` can synthesize fake AST elements for both good
and bad reasons (good = things like operator lowering; the type of
`x == y` is the same as `x.__eq__(y)` and we would ideally use it;
bad = things like CFG construction which can make utterly fake
ASt nodes, sometimes at the same locations as real nodes), for the
moment I'm doing the safe thing and only accepting the `typeCheck`
data if the expression matches; I fall back to naively using
the resolution if that fails.

To continue improving support we'll want to reduce the frequency
of fallback by making Pyre aware of the cases where it is safe to
use `typeCheck` results even though the expressions may differ.

The known behavior changes here:
- Operators will often come out to `Any` inside generator expressions, because
  `typeCheck.ml` is traversing a lowered form of the operator so the expressions
  don't match, and we fall back to a naive approach. Since the naive appraoch
  uses statement resolution, it will fail inside of generator expressions and
  we get a placeholder type like `Any`.
- There's a not-yet explained regression where a tuple type becomes unknown.
  The root cause is probably the same as for operator lowering, that somehow
  `typeCheck.ml` is saving a different expression for this location and so
  we give up and fall back to the naive appraoch, which doesn't know about
  generator-bound vars.
- Sometimes generator target types are slightly different; this isn't a real
  regression because these types are made-up anyway and mainly IDE
  facing.

I'm planning to address the first two regressions in upcoming diffs.

Reviewed By: grievejia

Differential Revision: D62449018

fbshipit-source-id: 1a9b555a7de39a2978a8377c5c3812f5c42a7572