Apply class intervals when propagating taint via TITO
Summary:
This diff adds class interval constraints when propagating taint via TITO.

Implementation-wise, this diff does the following:
1. Add class intervals to `CallInfo.Tito`.
2. Rename `Domains.transform_call_info` to `Domains.transform_tito`, since (1)
`CallInfo.Tito` is now parameterized by class intervals and hence we would
otherwise need to match `CallInfo.Tito` with any interval (2) all call sites to
this function uses `CallInfo.Tito`.
3. Update the return type of `Domains.apply_class_interval` so that it returns
an option type, which (1) eliminates the need of constructing an unused top class
interval, and (2) it could be directly used in `List.find_map` in
`Domains.apply_class_intervals_for_tito`.

Most changes in *.models are due to distinguishing CallInfo.Tito with
additional class intervals (but this is not shown in the *.models files in this
 diff -- this is shown in the next diff). By contrast, previsouly we merge all
LocalTaint under a single CallInfo.Tito.
The model changes in `class_interval.py` are due to eliminating the two false
positives (from the last diff).

Reviewed By: arthaud

Differential Revision: D62446861

fbshipit-source-id: 25d965c5c5569504e4e5d4c7612f3c2c7b4445bc