For functions, let the typechecker recognize PEP695 type vars, paramSpec and typeVarTuple as type annotations
Summary:
There are multiple stages for processing function defintions, which include `initial `and `forward_statement`. After `initial` is called, it is followed by a check which validates annotations and if an annotation is untracked, we raise an error.

This diff is to consider a type variable as an annotation. I'm doing so by collecting the type variable and inputting it into the resolution. By the time we check for untracked errors, if we see the variable in the resolution, we do not raise the error.  We do so at a stage where we visit type annotations, and do so for every name in the annotation by creating a variable out of it and checking it in the local resolution. This required threading the local resolution through the relavant parts of the code.

From looking at the code, this is not to be confused with `parse_annotation` parsing the type variable as "Any".

I initially I hardcoded the type variable into the result of `parse_annotation` and saw no effect on the type error we're trying to remove here. This makes sense though since the output of parse_annotation is "Any" and we can see that we're looking at a primitive as the type variable in question. So the result of `parse_annotation` isn't directly used to check for untracked annotation errors.

 It looks to me like we still have to allow `parse_annotation` to take scoped type variables from `initial`, where we create a map from `type_params` and pass it down to `parse_annotation`.

This will require a series of refactors since parse_annotation in GlobalResolution doesn't currently consider `scoped_type_vars`.

Reviewed By: stroxler

Differential Revision: D62527148

fbshipit-source-id: d72cfd543241ce6171421a76009343ca7f9c0fdb