Move decorator application to attribute instantiation
Summary:
Pyre stores attribute tables for each class, containing the "unannotated
attributes" for that class. These tables are used to resolve attributes for
expressions like `foo.bar`.

Attribute tables are cached using the standard environment abstraction, which
combines shared memory caching and dependency tracking. This abstraction is not
able to cache _recursive_ access, so if creating an attribute table requires
attribute tables, those nested accesses can not come from the cache.

Before this diff, unannotated attributes for methods were types that had all
decorators applied. However, applying decorators requires attribute tables to
look up the decorator. Because nested access to attribute tables can not be
cached, Pyre needed to create those attribute tables repeatedly.

This uncacheable behavior would blow up. For example, consider the following
code:

```
class C:
  foo.bar()
  baz() -> None: ...
```

To create the attribute table for `C`, we need the attribute table for `foo`.
If any of the methods in `foo` also have decorators, then all of those
decorators need to be applied as well, not just `bar`'s decorators. If any of
those decorators require attribute tables, then those attribute tables need to
be computed as well, and so on.

So the computation blows up, but also none of it can be cached.

This diff stops trying to apply decorators in unannotated attributes. Instead,
we store the undecorated method signature along with the decorator expressions.
We apply decorators when instantiating the attribute instead. This avoids the
blowup and improves cacheability.

There is a trade-off here: applying decorators is not always cached. Global
annotations are cached, but decorators for class methods are applied each time
we instantiate the attribute. However, applying decorators is fairly simple
compared to attribute instantiation overall. That said, it might make sense to
consider caching attribute instantiation.

Reviewed By: stroxler

Differential Revision: D62557000

fbshipit-source-id: 47a11f3dd15aa122e8a7f2c14eb161ed8b03660c