Remove parse annotation cache
Summary:
The parse annotation cache stores, as the name suggests, the type corresponding
to each annotation. Creating a type from an annotation should be cheap, so it's
not clear why we need this cache. Also, the introduction of PEP-695 scoped type
variables make cache invalidation more complicated, since scoped type variables
are not stored in the global environment.

This diff simply removes the cache. This depends on the previous diff, which
improves the cacheability of attribute tables. Without the previous diff,
removing the parse annotation cache is a significant performance regression,
because the blowup in attribute table computation also means we were parsing
annotations many many times.

Reviewed By: grievejia

Differential Revision: D61666321

fbshipit-source-id: b63421b99a5eed5e7e63e866f181370bf5ddc7ae