Rearrange attributeResolution: move resolve_define_undecorated up
Summary:
I realized somewhat belatedly that `resolve_define_undecorated` belongs near
the very bottom of the stack the way that I'm rearranging
`attributeResolution`, because it is basically at the same abstraction level
as `parse_annotation`: it just looks at the signature and lifts the expression
level to the type level, producing a bare callable type (and deferring decorator
handling for later).

This diff moves it up.

Reviewed By: grievejia

Differential Revision: D62755759

fbshipit-source-id: 3f65237fc26d9a4958d7b31c5723a632d8edcf32