Rearrange attributeResolution: move annotation parsing to top
Summary:
After reviewing the change to move decorator resolution out of the
uninstantiated attribute table logic last week, I decided that the time has
come to start making sense of attributeResolution.ml as a whole.

A first step is to arrange the methods in what seems like a more logical order;
while it's true that in today's code they all do form a mutually recursive
loop, there is an ordering that does make sense based on a few factors:

- First, `parse_annotation` only depends on the rest through `full_order`
  being used to validate type variable bounds; it's a refactor away from actually
  being cleanly separated as a lower layer. Hence this diff: move it to be at
  the top, emphasizing that it is the base layer.

- Second, the ordering of the cache tables does imply some amount of "layering"
  even though the logic is mutually recursive: if "lower" layers call into
  "upper" layers (which normally would only happen via `full_order` usage) then
  the calls are uncached. Having the order of code reflect the physical layers
  of the cache tables is useful.

This diff starts out by moving annotation validation and parsing to be at the
top, instead of the `get_typed_dictionary` method (which really belongs near
the bottom).

To keep diffs easy to review I'm going to rearrange the methods gradually, I'll
be stacking more diffs to continue reordering.

Reviewed By: grievejia

Differential Revision: D62755762

fbshipit-source-id: 7cda16428b8b0d2923be8aea89459f05f8185e3a