Rearrange attributeResolution: move metaclass up
Summary:
The `metaclass` method depends only on `parse_annotation` and
on `full_order` (which is the main source of mutual recursion
in `attributeResolution.ml`). As a result, there's a pretty
solid argument that it logically should be layered as one of the
"bottom" level methods, just above `parse_annotation`.

In addition, this matches the caching strategy: prior to the recent
elimination of ParseAnnotationCache the MetaclassCache was the bottom
cache layer, and now it is in fact at the very bottom!

Let's rearrange both the .ml and .mli file to reflect this ordering.

Reviewed By: grievejia

Differential Revision: D62755763

fbshipit-source-id: 92fb0c4672c0b5d70403709ad854278210ecf000