Make a TODO noting that infer shouldn't compute default values
Summary:
We had a bug report this week where a user was blocked using `pyre infer`
because of a libcst parser error.

The root cause was that `pyre infer` was spitting out a stub file that included
`lambda $parameter$x: $parameter$x.__pow__(5)` or something along those lines.

Pyre shouldn't really be trying to figure out default values at all, this is
a problematic thing for a type checker built around ast mangling to try to
do, and it isn't needed for valid stubs (we can just use `...` to indicate the
presence of a default value).

To unblock the user quickly I'm fixing the problem in the frontend which works
fine and avoids the complications of a three-phase deploy where we make a
breaking change to the backend/frontend serialization layer, but I wanted to
include this TODO in the backend so that we don't replicate the problem if/when
we rewrite the `pyre infer` logic.

Reviewed By: grievejia

Differential Revision: D62757591

fbshipit-source-id: bb766364b7ae8e05247eba4855655828bb2245a3