Have infer use `...` for default parameters
Summary:
A user reported that they were blocked applying type annotations to a
source code because of a libcst parsing error involving a `$`.

The root cause is indeed qualification: Pyre infer is spitting out a
mangled lambda `lambda $parameter$x: $parameter$x.__pow__(5)` in
the stub for a function that has a default value.

This is silly though; in stubs we can just use `...` to indicate a
default value. This commit modifies the frontend of `pyre infer`
to make this change: if the backend passes a default value, we
use `...` in the stubs.

It would be better in principle to fix this in the backend too,
but changing the serialization is complicated in configurations
where the frontend and backend aren't upgraded in lock step,
and given that we aren't actively working on `infer` I don't think
it's worth doing right now. I did add a `TODO` in the backend source
code in D62757591 so that if we ever rewrite the backend using the
existing code as a reference we won't replicate this mistake.

Reviewed By: grievejia

Differential Revision: D62757804

fbshipit-source-id: 02458d139ab9f0588243a7f011fe715ac1fd465e