Rename `validate_and_sanitize_type_arguments`, document
Summary:
When trying to understand how the method whose name I am changing
here fits into Pyre, it is important to understand that it is handling
two different operations:
- validating type arguments, and producing error information on issues
- eliminating said issues by falling back to gradual types

It makes some sense to handle these operations together (I've worked on
many, many codebases with a "validate and sanitize" operation) because
as a rule both are triggered by the same conditions and we don't want
to duplicate that logic.

But the two operations are not the same: the `validate` part is used only
in a scenario like `typeCheck` where we can actually produce user-facing
error information, whereas the `sanitize` portion is used unconditionally
by `parse_annotation`.

In addition to the rename, I added a doc comment and explained where the
two operations are relevant.

Reviewed By: grievejia

Differential Revision: D62768107

fbshipit-source-id: 1286ff688401ae8d9ba82e5c2b8eddd9cc3f3df7