AnnotatedAttribute.create{_instantiated}
Summary:
The flow for an attribute is to first create an uninstantiated
attribute, and then depending on the use case convert maybe it to an
instantiated attribute.

Using `AnnotatedAttribute.create` to refer only to the instantiated
case is confusing to me; it makes the names of our operations inconsistent
(and I believe consistency is a big part of being able to navigate and
internalize an architecture), and also make it seem like the instantiated
version is more "canonical" than the uninstantiated version when in reality both
are useful at times.

Reviewed By: grievejia

Differential Revision: D62885251

fbshipit-source-id: bb0d454604960b8a8381efe251326202fcbcadd8