Namespace `constraints` to `Testing`
Summary:
I just realized, as I was preparing to write a doc comment for the
`constraints` method, that it is actually a private implementation
detail.

It gets leaked all the way to GlobalResolution's public interface
because OCaml's testing model requires this, but there is actually
*exactly one* callsite, in `instantiate_annotation`.

I considered just deleting it from the interfaces, which I think
might be the right move (at some point, too many tests can just
be dead weight), but given that this is subjective I decided to
keep it and just create a `Testing` namespace for this kind of
private hook that is only in the interface for testing.

Reviewed By: grievejia

Differential Revision: D62899929

fbshipit-source-id: 065b04465ec729baa6f9ab335778bdc43e38b6a4