Eliminate `?arguments` from `constraints` method
Summary:
I was reading the code, and couldn't figure out how it made any sense at all,
it seems to have confusing parameters with arguments baked into it.

Then I realized that except for a couple of unit tests, the `arguments` input
is never actually used! My conclusion is that the existing code indeed doesn't
make sense and we should just simplify. I have to do this now, because I would
be unable to write a coherent doc comment about the pre-existing behavior.

Reviewed By: grievejia

Differential Revision: D62899926

fbshipit-source-id: 51676eff5e4858bffe610e826205336dcfb2a2ee