Clean up telemetry around GLTC type checking
Summary:
In the previous telemetry that we were doing to determine the viability of PTT in the IDE, I added any telemetry that seemed useful, regardless of whether something similar existed or if the data would actually be useful. Now that we're preparing to roll out, I'm cleaning up the telemetry we're collecting.

The goal with this is to deduplicate data, remove anything that's not useful, and make sure the resulting telemetry dict cleanly separates data into different sections.

This diff organizes the telemetry we've been collecting around GLTC and cleans up the GLTC dataclass.

Reviewed By: stroxler

Differential Revision: D62783677

fbshipit-source-id: 9869a855e75a6adf14c50fdff82ff7f181cd83b2