Clean up PTT telemetry
Summary:
In the previous telemetry that we were doing to determine the viability of PTT in the IDE, I added any telemetry that seemed useful, regardless of whether something similar existed or if the data would actually be useful. Now that we're preparing to roll out, I'm cleaning up the telemetry we're collecting.

The goal with this is to deduplicate data, remove anything that's not useful, and make sure the resulting telemetry dict cleanly separates data into different sections.

This diff focuses on cleaning up telemetry around PTT, updating the telemetry dataclass and some of the functionality around it.

Reviewed By: grievejia

Differential Revision: D62884424

fbshipit-source-id: 929c28d655e01d55d2857ffade63dfbdf6d8ea53