Use PTT type errors in IDE telemetry
Summary:
In this diff, we switch to using an argfile for querying Buck, and make a few final changes to the telemetry dict.

In the telemetry dict, the `type_errors` key is removed, since it's something I added a while ago for debugging purposes. There aren't any metrics really built around that key, and type errors have been moved into the `type_check_metadata` key related under the specific source that produced them (i.e. we don't combine all the type errors, they're now present under GLTC or PTT section, depending on which source produced them).

The `error_message` key is used for some metrics, so it makes sense to keep that for now. In followup diffs, I'll adjust that functionality in dataswarm to look under the specific type check metadata section like we do for type errors.

Reviewed By: grievejia

Differential Revision: D62885215

fbshipit-source-id: 0448f08b5acada20763209cdbda76699542ac0bb