Pull typing query logic into separate method
Summary:
The typing query is a Buck BXL script run to determine if a given file should be type checked with PTT or GLTC.

While doing testing around PTT in the IDE, the typing query was bundled into the same method as the PTT call to keep experimental functionality together and to make the implementation/measurement easier.

Now that we're actually implementing PTT, this query will need to be done before kicking off the asynchronous GLTC and PTT type check coroutines. This diff pulls out that functionality and creates a new dataclass for holding its results.

Reviewed By: grievejia

Differential Revision: D62781505

fbshipit-source-id: 12cbd732ba6b36b71e78fb374026ff58116bc5d9