Rename Type.is_meta -> Type.is_builtins_type + improve some comments / varnames
Summary:
The name `Type.is_meta` sounds like it's asking whether the
current type is a metaclass (which isn't really a valid question,
since that's contextual, but loosely corresponds to whether the current
type inherits from `type`).

But that's not what it's doing, it's really just asking whether this
type *is* exactly `type[Xyz]` for some `Xyz`.

I came across this in the middle of trying to understand
`Type.class_data_for_attribute_lookup`, so I also do some commenting
and gentle renaming of a few things there.

Reviewed By: grievejia

Differential Revision: D62970311

fbshipit-source-id: 8e76df7bb75a2254209e03cc9811bd76d617796f