Rename class_data_for_attribute_lookup
Summary:
The name is a little confusing to me, it sounds like it is "data about a class
for attribute lookup", but that's not what it is, it's "data about a possible
lookup of an attribute on a class"; the ordering matters.

In addition, I decided to give the the function that computes a list of these
class-attribute-lookup a name that emphasizes that it's getting a list of these
for a given type.

Reviewed By: grievejia

Differential Revision: D62970314

fbshipit-source-id: 132581e5df6c6bd7b33d3ffeff965d24533f71cf