Rename `instantiated` -> `type_for_lookup`
Summary:
This diff renames `instantiated` in cases where it refers to the
type on which we are looking up an attribute to `type_for_lookup`.

This is a bit of a large rename, but I think it is necessary because
attribute lookups always have to deal with the question of whether
we are looking up uninstantiated vs instantiated attributes. As a
result, the term `instantiated` generally would be interpreted to mean
the attribute is instantiated, and using that variable name to represent
the lookup type in such code (especially in the `instantiated_attribute` method
of AttributeResolution!!) is pretty confusing.

Reviewed By: grievejia

Differential Revision: D62970312

fbshipit-source-id: a780e9856a56a8e370b8b65fe136ac30a6eaf7eb