New doc comment on `resolve_literal`
Summary:
The old comment didn't seem very descriptive to me - it seemed to be
explaining that we can't resolve arbitrary expressions, which is fine
but isn't really explaining what this function *does* do or how it is
used.

The new comment explains roughly what the function does, how it maps
to Pyre's user-facing behavior, and clearly states that we get
`Type.Any` as a fallback which I think is important to have in the
doc comment.

Reviewed By: grievejia

Differential Revision: D62973448

fbshipit-source-id: 65636db9a8129ec43318e9a7ab2d38fce2170be4