+1 fix named tuple subclass behavior
Summary:
The typing spec states:

> A named tuple class can be subclassed, but any fields added by the subclass are not considered part of the named tuple type.

In other words: if you have a named tuple class w/ 2 fields and make a subclass that adds a third field, that subclass behaves as if it was still a 2-tuple (though the new field may be accessed by name).

This actually works out well with our current preprocessing, since it only generates the tuple fields for classes which directly extend NamedTuple (so subclasses would only take the info of the parent class).

We needed to make a few tweaks to realize the behavior, however:
- Previously, we only checked for the `_fields` attribute defined in the current class. For subclasses of named tuple classes, we need to check the parents.
- For subclasses of named tuple classes, when we get the tuple field by name we need to set `~transitive:true` since the field is defined in a parent class.

With those small changes the feature should work as expected.

Note: It's unclear to me which class takes precedence when you do multiple inheritance and take the subclass of multiple named tuple classes. I just know the fields do not get merged:

```
>>> class Foo1(NamedTuple):
...     x: int
...
>>> class Foo2(NamedTuple):
...     y: int
...
>>> class Foo3(Foo1, Foo2):
...     z: int
...
>>> Foo3._fields
('x',)
```

Reviewed By: stroxler

Differential Revision: D62890746

fbshipit-source-id: 2d13d946c1742ed22090881fba074d52922d33e5