Fix PyreReadOnly bug involving positional-only parameters
Summary: I noticed while playing around with PyreReadOnly that mutating method violations weren't being reliably reported. As it turns out, the violation isn't reported when a method uses the new positional-only parameter syntax, because the logic for type-annotating `self` expects `self` to be a named parameter. This bug has probably been around for a long time, but it is much more noticeable now because of the typeshed upgrade I pulled in that applies the positional-only parameter syntax to all of the stdlib.

Reviewed By: yangdanny97

Differential Revision: D63080126

fbshipit-source-id: a969d57224d96df410d05eb456bde3b825921968