Fix a bug in typecheck where a type variable is not interpreted as a type
Summary:
In this diff stack, we observed an error in generic classes of the form "annotation T is not defined as a type". This error only existed on bounds.

The reason this happened is that in typecheck, the existing logic checks annotations in resolution by name only.  We're specifically referring to this line:

    `let add_untracked_errors errors =
          Type.collect_names annotation...`

I don't think that at this stage that we need to consider bounds in the equality check. So I made another version of it which only compares by name, which fixes the error.

Reviewed By: stroxler

Differential Revision: D63077918

fbshipit-source-id: 3c488d2f35a67ee9d49f7fd419142a4a3eb475cf