Support bounds for generic functions
Summary:
We handle bounds for PEP695 syntax by deciding if they are Explicit or Bound and transforming them into types.

The following is not allowed:

```
class A[U]:
    def func[T: U](self, a: T) -> T:
        return a
```

Therefore, we do not have to resolve variables and we pass an empty variable map.

In the next diffs, we apply the same logic to type statements and generic classes

Reviewed By: stroxler

Differential Revision: D63058463

fbshipit-source-id: 9bb70a0ca21c741acf93da88abf3d5fd0d1f7d94