Add a unit test that shows that assert type does not recognize ParamSpec but recognizes TypeVar and TypeVarTuple for generic classes
Summary:
As I was going through conformance tests, I noticed the following testcase, which I added as a unittest.

We are recognizing TypeVarTuple and TypeVar, but not ParamSpec.

This is a bit strange, since it's only ParamSpec that's not recognized. I checked the different transformations in attributeResolution, preprocessing, typecheck and classHierarchyEnvironment in case I accidently missed a transformation, but they are correct.

So this will need some further investigation.

Reviewed By: samwgoldman

Differential Revision: D63142551

fbshipit-source-id: c938822d14331d3a8b499f814a80caf08158574d