+1 disallow named tuple fields without defaults from appearing after fields with defaults
Summary:
Since the init parameters for named tuples are generated in order of declaration, we can't have fields without defaults appearing after fields with defaults.

Our current preprocessing step makes default and non-default fields indistinguishable when looking at the field definition in the class (the default values are only represented in the parameter of the generated new/init functions).

This diff modifies the preprocessing pass to avoid inserting `= ...` for fields that do not have default values.

Eventually, the preprocessing will be removed entirely, but the typechecking logic for detecting default vs non-default fields should stay roughly the same (checking for explicitly assigned values in the field definition).

Reviewed By: stroxler

Differential Revision: D62975709

fbshipit-source-id: ed9a7fef2d67b05ff77e36f5447b8a8c66e5c453