Lay the ground work for bound validation in typecheck
Summary:
We add a new type error for bound validation for the new generic syntax, lay the groundwork for emitting the error and add a unittest.

What remains is the actual check, where we validate if the expression has the right format.

If this groundwork looks good, then the next diff is to add the validation where the TODO comment is.

Reviewed By: stroxler

Differential Revision: D63155961

fbshipit-source-id: b7b8978b319e1bf00361aa487aeedd914e132e06