Add cwd to Buck root query error message
Summary:
The Pyautotargets buck root query keeps failing for some reason, but I'm not sure why. This diff adds the cwd so that we can see where the buck subprocess call is happening from.

My current theory is that eden isn't running, but I'm honestly not too sure about that.

Reviewed By: aleivag

Differential Revision: D63291969

fbshipit-source-id: 3bd9e4c97187a85816d424d08bbcc4be6e0feded