Rename PyrePysaApi -> PyrePysaEnvironment
Summary:
We decided at the last weekly sync that it would be a good idea to
isolate not only Pysa access to shared-memory-backed environment data,
but also to pure logic in Pyre.

One issue that comes up is that when I do this, the name of the existing
api module becomes less clear, because "Api" doesn't clearly suggest that
this is specifically restricting access to symbol tables as opposed to
access to pure functions for operating on types and such.

So I decided to rename first... PyrePysaEnvironment seems like it
pretty clearly has connotations of involving access to the "environment"
table data as opposed to other aspects of Pyre analysis.

Reviewed By: arthaud

Differential Revision: D63274473

fbshipit-source-id: 3edce5a5419bda108ce8c053c48b64ecbc3149cd