Start bootstrapping PyrePysaLogic: cut over ForwardAnalysis
Summary:
**Stack description**

This diff kicks off a stack where I move forward on a plan the Pyre and
Pysa team came up with to make the coupling between the type system and
Pysa more explicit: I am going to eliminate all direct uses of the `Analysis`
module from Pysa code module, and instead thread all such uses through a
`PyrePysaLogic` module.

The idea is just like the earlier PyrePysaEnvironment refactor, which made
all Pysa access to Pyre's *environment* (as in the actual symbol tables)
go through a single API.

This doesn't actually change anything functionally, but the existence of a
consolidated module with all of the couplings has a several benefits for
planning:
- we can see in a single place what in Pyre is used by Pysa
- it's much easier to search for callsites and figure out *how* a specific
  function is used when it is always namespaced behind a globally unique
  name like `PyrePysaLogic`

--------

**Diff description**

I decided the easiest way to do this is to just go through Pysa modules one
by one cutting over all direct uses of `Analysis`.

This is just the first such diff, which cuts over `ForwardAnalysis.ml`

Reviewed By: arthaud

Differential Revision: D63283397

fbshipit-source-id: 48b16b5fe07c54bf2e3a35a292e329e1f62545ae