Emit error on attempt to write to read-only TypedDict field
Summary:
Modifies the `__setitem__` overloads for TypedDict so that calling `__setitem__` with a read-only key always generates an error.

A follow-up change will improve the error message.

Note on interaction with PEP 728 (TypedDict with Typed Extra Items):

This approach will break if we implement `__extra_items__` by adding a fallback `__setitem__` overload. However, this problem is not limited to `ReadOnly`; a fallback overload could also match if a user attempts to set a regular read-write key to a value of the wrong type. Given that this problem is not ReadOnly-specific and PEP 728 will be in Python 3.14 at the earliest, I think it's okay to punt on it for now.

Reviewed By: stroxler

Differential Revision: D63338634

fbshipit-source-id: 74723cfb6bee562a0321cb9d8dace7285b7be44a