Fix bound creation and bound check by qualifying type parameters in preprocessing
Summary:
Before, we were not qualifying type parameter bounds in preprocessing, which caused type creation to be incorrect, and we relised on this for bound creation and validation.

This caused bounds that weren't qualified to be "any" which did not raise the right type errors.

Additionally, I relied on being able to parse an annotation, and getting "any" made that check faulty,

In this diff, we fix these problems by qualifying bounds (which are expressions) in preprocessing.

The extra "constant" check is not needed since we parse strings as a type as well if they are valid.

Reviewed By: stroxler

Differential Revision: D63354077

fbshipit-source-id: 5058f2e47fa26eb976ac909d00a2d81c55a16808