+1 improve `sys.platform` checks
Summary:
Previously, the only preprocessing we did was statically removing branches guarded by `sys.platform == "win32"`.

This diff adds support for more robust checks against other platforms, and also for `sys.platform.startswith` checks, which are recommended by the official documentation.

The platform string will be passed from the pyre client, and will be plumbed through in future diffs. For now the default is linux.

There is no support for complex and/or/not conditions or membership checks in a list, but that can be added if there's demand in the future.

Reviewed By: stroxler

Differential Revision: D63336519

fbshipit-source-id: d58b33227b6e6560dc2dd2c435d303eddd128ac4