Recognize PEP695 ParamSpec as types
Summary:
We rely on typeAliasEnvironment/global scope to define paramSpec types for typechecking.

This is seperate from attributeResolution where we resolve types, as shown by the unittest in the previous stack.

In this diff, we recognize paramSpec types, by looking at the local resolution first before looking at the global resolution.

Reviewed By: stroxler

Differential Revision: D63430155

fbshipit-source-id: df77e33abd49383b294f928e997536e3bed7cc54