parse subscript expression as a type if base is a type
Summary:
See the linked task for additional context/examples. Right now, when expressions representing generic types occur in places where we don't expect annotations, we treat them as regular subscripts and check it as a `__getitem__` call. This results in strange errors like `list has no attribute __getitem__` when typechecking an expression like `list[str]`.

This diff changes it so that if the base of the subscript resolves to a `type[]`, we attempt to parse the whole subscript expression as a type annotation. If the expression parses into a non-Top type with no errors, then we use that. Otherwise, we'll fall back to the old pattern with `__getitem__`.

In the future, we can revisit what happens if it parses into a type but has errors, but I think that tweak is easier to make as an incremental change once this is landed.

Another note is that it may not always be correct to parse these expressions as types, because the base type's metaclass could actually implement a `__getitem__` method, which would allow it to have custom behavior in subscripts. This seems like an uncommon use case though.

Reviewed By: stroxler

Differential Revision: D63402954

fbshipit-source-id: af3c9e43e81c0dd4de13673936e3269ef7d39e0e