Simplify: class hierarchy does not need to be leaked
Summary:
The main goal of collecting Pyre logic used by Pysa in PyrePysaLogic is to
make it easy to see what all that logic is (all of it is potentially
exposed if we try to re-negotiate the coupling using typed asts).

So it makes sense to simplify where we can. We were exposing the entire
class hierarchy, which is a *very* internal implementation detail of Pyre,
but the only reason we were doing it was to get immediate parents, which is
already in the PyrePysaEnvironment api anyway and is far less invasive.

Reviewed By: rchen152

Differential Revision: D63471243

fbshipit-source-id: eda78a2d0317d79cb78790b5bcfb137cb132367e