Begin plumbing for variance inference
Summary:
We add a new variance that represents pre variance inference. We change generic parameters and variable declarations to due the preinference variance.

We add a function in attribute resolution which goes from pre variance inference to post variance inference
Type.ml needs more work. We currently convert pre variance to post variance in place. What we need to do is to not reason about variance in Type.ml, but in the constraint resolution stage.
There are two places where this is happening which are maked by TODOs.

Next, I also have to consider some kind of environment that goes from name to variance. I am still thinking about where this will go, but it looks like we could use it inplace of `look_up_variance`

Reviewed By: stroxler

Differential Revision: D63348113

fbshipit-source-id: 8467954c7d6398e84e1f2ee694287982acd722bc