Finish plubing for variance inference
Summary:
1- We expose our variable map interface from attribute resolution
2- We use it for all occurrences in the code where we need variance, such as constraint resolution

Some additional notes:
We default to invariant when we cannot infer anything
I'm not using the class name currently as I just read variance off the parameter so my class names could be wrong. I think it should mostly be accurate though since I use them to get the class parameters and all unittests pass.

We also flip the "Invariant" in classHierarchyEnvironment to "Undefined". This would be flipped back to  "Invariant" through our map, so we shouldn't see any change in conformance.

Reviewed By: stroxler

Differential Revision: D63421327

fbshipit-source-id: 8e9ad28937268f44647ad0a2bda7aec7d91339cc