Add support for pyre_extensions.PyreReadOnly being the primary name
Summary:
With this change, pyre supports either of pyre_extensions.ReadOnly or pyre_extensions.PyreReadOnly being the primary (non-aliased) name for the PyreReadOnly object.

After this, we can switch around the definitions in pyre-extensions, then delete support for ReadOnly as the primary name, then delete the pyre_extensions.ReadOnly alias.

Reviewed By: stroxler

Differential Revision: D63501515

fbshipit-source-id: 8d585f1d70170146c3e74e21cc06228f2e23d775