Move a lot of model logic to PyrePysaEnvironment
Summary:
The ModelVerifier codebase had a lot of logic that wasn't actually about
verifying models at all, it was about running custom queries against Pyre with
weird semantics (in particular the `resolve_qualified_name_to_global` semantics
are extremely specific to the Pysa DSL)

This is exactly the kind of thing that cannot continue to live on the Pysa side,
it needs to be jointly maintained as we consider changes to Pyre's environment,
because the fact that the semantics are a bit werid means a lot of pyre's
implementation details are relevant to how this code works.

Any major changes to Pyre, and we woudl need to work out exactly what semantics
are truly needed and rewrite this code from scratch.

Reviewed By: arthaud

Differential Revision: D63498403

fbshipit-source-id: 2fd11fd100586ff54f44daf64f41797c3d9d7bf8