Cleanup: move tests of resolve_qualified_name_to_global
Summary:
I didn't want to muck up the previous diff - which was more substantial
than most of my Pyre/Pysa refactors - with unimportant test changes.

But the fact is that this test was entirely of the function we moved into
pyrePysaEnvironment, so the test ought to live there going forward.

Reviewed By: arthaud

Differential Revision: D63498405

fbshipit-source-id: bd4e388d8c0191296e6719c779e63d8cf0ac0a29